requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
bcrypt==5.0.0
boto3==1.42.71
botocore==1.42.71
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==5.2.8
dnspython==2.8.0
edge-tts==7.2.3
frozenlist==1.8.0
idna==3.11
invoke==2.2.1
jmespath==1.1.0
mollie-api-python==3.9.1
multidict==6.7.0
oauthlib==3.3.1
paramiko==4.0.0
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pymongo==4.15.4
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
sshtunnel==0.4.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
yarl==1.22.0
gunicorn==23.0.0
